requirements.txt
Pillow===9.2.0
pytweening===1.0.4-py3.10.egg-info
PyScreeze===0.1.28
PyRect===0.2.0-py3.10.egg-info
pyperclip===1.8.2
PyMsgBox===1.0.9
PyGetWindow===0.0.9-py3.10.egg-info
MouseInfo===0.1.3-py3.10.egg-info
PyAutoGUI===0.9.53-py3.10.egg-info
sortedcontainers===2.4.0
urllib3===1.26.11
sniffio===1.2.0
PySocks===1.7.1
pycparser===2.21
idna===3.3
h11===0.13.0
certifi===2022.6.15
attrs===22.1.0
async_generator===1.10
wsproto===1.1.0
outcome===1.2.0
cffi===1.15.1
trio===0.21.0
cryptography===37.0.4
trio_websocket===0.9.2
pyOpenSSL===22.0.0
selenium===4.3.0